ref(MungeLocalSdp): move to RTC module
